chore: leverage nasconvert lib
