+++ internal/amf/nas/dispatch.go
@@ -4,115 +4,3 @@
 // SPDX-License-Identifier: Apache-2.0
 
 package nas
-
-import (
-	ctxt "context"
-	"errors"
-	"fmt"
-
-	"github.com/ellanetworks/core/internal/amf/context"
-	"github.com/ellanetworks/core/internal/amf/nas/gmm"
-	"github.com/ellanetworks/core/internal/logger"
-	"github.com/free5gc/nas"
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/trace"
-	"go.uber.org/zap"
-)
-
-var tracer = otel.Tracer("ella-core/amf/nas")
-
-func Dispatch(ctx ctxt.Context, ue *context.AmfUe, msg *nas.Message) error {
-	if msg.GmmMessage == nil {
-		return errors.New("gmm message is nil")
-	}
-
-	if msg.GsmMessage != nil {
-		return errors.New("gsm message is not nil")
-	}
-
-	if ue.State == nil {
-		return fmt.Errorf("ue state is empty")
-	}
-
-	msgTypeName := MessageName(msg.GmmMessage.GmmHeader.GetMessageType())
-	spanName := fmt.Sprintf("AMF NAS %s", msgTypeName)
-
-	ctx, span := tracer.Start(ctx, spanName,
-		trace.WithAttributes(
-			attribute.String("nas.messageType", msgTypeName),
-		),
-	)
-	defer span.End()
-
-	logger.AmfLog.Info(
-		"Received NAS message",
-		zap.String("MessageType", msgTypeName),
-		zap.String("SUPI", ue.Supi),
-	)
-
-	return gmm.HandleGmmMessage(ctx, ue, msg.GmmMessage)
-}
-
-func MessageName(code uint8) string {
-	switch code {
-	case nas.MsgTypeRegistrationRequest:
-		return "RegistrationRequest"
-	case nas.MsgTypeRegistrationAccept:
-		return "RegistrationAccept"
-	case nas.MsgTypeRegistrationComplete:
-		return "RegistrationComplete"
-	case nas.MsgTypeRegistrationReject:
-		return "RegistrationReject"
-	case nas.MsgTypeDeregistrationRequestUEOriginatingDeregistration:
-		return "DeregistrationRequestUEOriginatingDeregistration"
-	case nas.MsgTypeDeregistrationAcceptUEOriginatingDeregistration:
-		return "DeregistrationAcceptUEOriginatingDeregistration"
-	case nas.MsgTypeDeregistrationRequestUETerminatedDeregistration:
-		return "DeregistrationRequestUETerminatedDeregistration"
-	case nas.MsgTypeDeregistrationAcceptUETerminatedDeregistration:
-		return "DeregistrationAcceptUETerminatedDeregistration"
-	case nas.MsgTypeServiceRequest:
-		return "ServiceRequest"
-	case nas.MsgTypeServiceReject:
-		return "ServiceReject"
-	case nas.MsgTypeServiceAccept:
-		return "ServiceAccept"
-	case nas.MsgTypeConfigurationUpdateCommand:
-		return "ConfigurationUpdateCommand"
-	case nas.MsgTypeConfigurationUpdateComplete:
-		return "ConfigurationUpdateComplete"
-	case nas.MsgTypeAuthenticationRequest:
-		return "AuthenticationRequest"
-	case nas.MsgTypeAuthenticationResponse:
-		return "AuthenticationResponse"
-	case nas.MsgTypeAuthenticationReject:
-		return "AuthenticationReject"
-	case nas.MsgTypeAuthenticationFailure:
-		return "AuthenticationFailure"
-	case nas.MsgTypeAuthenticationResult:
-		return "AuthenticationResult"
-	case nas.MsgTypeIdentityRequest:
-		return "IdentityRequest"
-	case nas.MsgTypeIdentityResponse:
-		return "IdentityResponse"
-	case nas.MsgTypeSecurityModeCommand:
-		return "SecurityModeCommand"
-	case nas.MsgTypeSecurityModeComplete:
-		return "SecurityModeComplete"
-	case nas.MsgTypeSecurityModeReject:
-		return "SecurityModeReject"
-	case nas.MsgTypeStatus5GMM:
-		return "Status5GMM"
-	case nas.MsgTypeNotification:
-		return "Notification"
-	case nas.MsgTypeNotificationResponse:
-		return "NotificationResponse"
-	case nas.MsgTypeULNASTransport:
-		return "ULNASTransport"
-	case nas.MsgTypeDLNASTransport:
-		return "DLNASTransport"
-	default:
-		return fmt.Sprintf("Unknown message type: %d", code)
-	}
-}
