+++ internal/amf/util/nas_plmn_id_convert.go
@@ -1,7 +1,6 @@
 package util
 
 import (
-	"encoding/hex"
 	"fmt"
 	"strconv"
 
@@ -57,22 +56,4 @@ func PlmnIDToNas(plmnID models.PlmnID) ([]uint8, error) {
 	}
 
 	return plmnNas, nil
-}
-
-func PlmnIDToString(nasBuf []byte) string {
-	mccDigit1 := nasBuf[0] & 0x0f
-	mccDigit2 := (nasBuf[0] & 0xf0) >> 4
-	mccDigit3 := (nasBuf[1] & 0x0f)
-
-	mncDigit1 := (nasBuf[2] & 0x0f)
-	mncDigit2 := (nasBuf[2] & 0xf0) >> 4
-	mncDigit3 := (nasBuf[1] & 0xf0) >> 4
-
-	tmpBytes := []byte{(mccDigit1 << 4) | mccDigit2, (mccDigit3 << 4) | mncDigit1, (mncDigit2 << 4) | mncDigit3}
-
-	plmnID := hex.EncodeToString(tmpBytes)
-	if plmnID[5] == 'f' {
-		plmnID = plmnID[:5] // get plmnID[0~4]
-	}
-	return plmnID
 }
