+++ internal/amf/util/ngap_plmn_id_convert.go
@@ -26,17 +26,21 @@ func PlmnIDToNgap(modelsPlmnid models.PlmnID) (*ngapType.PLMNIdentity, error) {
 	var hexString string
 	mcc := strings.Split(modelsPlmnid.Mcc, "")
 	mnc := strings.Split(modelsPlmnid.Mnc, "")
+
 	if len(modelsPlmnid.Mnc) == 2 {
 		hexString = mcc[1] + mcc[0] + "f" + mcc[2] + mnc[1] + mnc[0]
 	} else {
 		hexString = mcc[1] + mcc[0] + mnc[0] + mcc[2] + mnc[2] + mnc[1]
 	}
 
-	var ngapPlmnID ngapType.PLMNIdentity
 	plmnID, err := hex.DecodeString(hexString)
 	if err != nil {
 		return nil, fmt.Errorf("error decoding hex string: %s", err)
 	}
+
+	var ngapPlmnID ngapType.PLMNIdentity
+
 	ngapPlmnID.Value = plmnID
+
 	return &ngapPlmnID, nil
 }
