+++ internal/amf/util/ngap_ran_id.go
@@ -2,6 +2,7 @@ package util
 
 import (
 	"github.com/ellanetworks/core/internal/models"
+	"github.com/free5gc/ngap/ngapConvert"
 	"github.com/free5gc/ngap/ngapType"
 )
 
@@ -18,7 +19,7 @@ func RanIDToModels(ranNodeID ngapType.GlobalRANNodeID) models.GlobalRanNodeID {
 		if ngapGnbID.GNBID.Present == ngapType.GNBIDPresentGNBID {
 			choiceGnbID := ngapGnbID.GNBID.GNBID
 			gnbID.BitLength = int32(choiceGnbID.BitLength)
-			gnbID.GNBValue = BitStringToHex(choiceGnbID)
+			gnbID.GNBValue = ngapConvert.BitStringToHex(choiceGnbID)
 		}
 	case ngapType.GlobalRANNodeIDPresentGlobalNgENBID:
 		ngapNgENBID := ranNodeID.GlobalNgENBID
@@ -27,13 +28,13 @@ func RanIDToModels(ranNodeID ngapType.GlobalRANNodeID) models.GlobalRanNodeID {
 		switch ngapNgENBID.NgENBID.Present {
 		case ngapType.NgENBIDPresentMacroNgENBID:
 			macroNgENBID := ngapNgENBID.NgENBID.MacroNgENBID
-			ranID.NgeNbID = "MacroNGeNB-" + BitStringToHex(macroNgENBID)
+			ranID.NgeNbID = "MacroNGeNB-" + ngapConvert.BitStringToHex(macroNgENBID)
 		case ngapType.NgENBIDPresentShortMacroNgENBID:
 			shortMacroNgENBID := ngapNgENBID.NgENBID.ShortMacroNgENBID
-			ranID.NgeNbID = "SMacroNGeNB-" + BitStringToHex(shortMacroNgENBID)
+			ranID.NgeNbID = "SMacroNGeNB-" + ngapConvert.BitStringToHex(shortMacroNgENBID)
 		case ngapType.NgENBIDPresentLongMacroNgENBID:
 			longMacroNgENBID := ngapNgENBID.NgENBID.LongMacroNgENBID
-			ranID.NgeNbID = "LMacroNGeNB-" + BitStringToHex(longMacroNgENBID)
+			ranID.NgeNbID = "LMacroNGeNB-" + ngapConvert.BitStringToHex(longMacroNgENBID)
 		}
 	case ngapType.GlobalRANNodeIDPresentGlobalN3IWFID:
 		ngapN3IWFID := ranNodeID.GlobalN3IWFID
@@ -41,7 +42,7 @@ func RanIDToModels(ranNodeID ngapType.GlobalRANNodeID) models.GlobalRanNodeID {
 		ranID.PlmnID = &plmnid
 		if ngapN3IWFID.N3IWFID.Present == ngapType.N3IWFIDPresentN3IWFID {
 			choiceN3IWFID := ngapN3IWFID.N3IWFID.N3IWFID
-			ranID.N3IwfID = BitStringToHex(choiceN3IWFID)
+			ranID.N3IwfID = ngapConvert.BitStringToHex(choiceN3IWFID)
 		}
 	}
