--- internal/amf/util/service_area_list_convert.go
@@ -1,1 +1,0 @@
-package util
