+++ internal/decoder/nas/identity_response.go
@@ -3,7 +3,6 @@ package nas
 import (
 	"encoding/hex"
 
-	"github.com/ellanetworks/core/internal/amf/util"
 	"github.com/ellanetworks/core/internal/decoder/utils"
 	"github.com/free5gc/nas/nasConvert"
 	"github.com/free5gc/nas/nasMessage"
@@ -55,7 +54,7 @@ func buildMobileIdentity(mobileIdentity5GS nasType.MobileIdentity) MobileIdentit
 			PLMNID:   &plmnIDModel,
 		}
 	case nasMessage.MobileIdentity5GSType5gGuti:
-		_, guti := util.GutiToString(mobileIdentity5GSContents)
+		_, guti := nasConvert.GutiToString(mobileIdentity5GSContents)
 		return MobileIdentity{
 			GUTI:     &guti,
 			Identity: utils.MakeEnum(identityTypeUsedForRegistration, "5G-GUTI", false),
